requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-google-genai
langgraph
fastembed
tavily-python
python-dotenv
pydantic
rich
pytest
streamlit
requests
sentence-transformers
faiss-cpu
